% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_help.R
\name{get_dataset_help}
\alias{get_dataset_help}
\title{Extraire un contexte depuis l'aide d'un dataset}
\usage{
get_dataset_help(
  data_name,
  package = NULL,
  sections = c("^Description", "^Format")
)
}
\arguments{
\item{data_name}{data.frame ou nom du dataset}

\item{package}{Nom du package (chaîne), optionnel}

\item{sections}{Sections à extraire (regex), ex. c("^Description", "^Format")}
}
\value{
Chaîne avec le contexte ; NULL si non trouvé.
}
\description{
Tente de récupérer le texte d'aide associé à un jeu de données.
- `data_name` peut être un objet `data.frame` ou un nom de dataset (symbole ou chaîne).
- Si `package` est NULL, on essaie de retrouver automatiquement le package d'origine.
- On renvoie les sections les plus utiles (Description, Format) par défaut.
}
